Fix the four things the first CI run found, none of which any local run could
The repository had no git remote until today, so no workflow had ever executed. All four jobs failed
on their first attempt, and three of the causes were pre-existing bugs that were unreachable from a
developer's machine.

A PLATFORM GUARD WRITTEN AS AN ASSERTION. `vault.test.ts` had a case named "does not consult the
Linux-only backend on Windows or macOS" whose first line was
`expect(process.platform).not.toBe("linux")`. That is a guard expressed as a check: on a Linux
runner it is a FAILING TEST rather than a skipped one — `expected 'linux' not to be 'linux'` — and it
took down the desktop suite on all three operating systems. `it.skipIf` says the same thing in the
place that can act on it.

A `scripts` PACKAGE COLLISION THAT DEPENDED ON THE WORKING DIRECTORY.
`test_mint_desktop_session.py` did `from scripts import mint_desktop_session`, and there is a
`scripts` package at the repository root as well as under `apps/api`. From `apps/api` it resolved;
from the root it did not, and CI runs `pytest apps/api/tests` from the root — so the API job exited 2
at collection without running a single test. It loads by PATH now, the way `test_tunnel.py` already
did. I noted this hazard in a commit message a few hours ago; it is a bug, not a note.

`localhost` IN ci.yml, AGAINST ITS OWN SUITE'S ADVICE. `conftest.py` spends a paragraph on why the
database address must be `127.0.0.1`: on a dual-stack host `localhost` resolves to both 127.0.0.1
and ::1, Docker publishes an IPv4 listener only, and the reachability probe requires EVERY resolved
address to answer. Harmless while that meant a skip. Not harmless now that `REQUIRE_POSTGRES=1`
turns an unreachable database into a collection-time failure — which is the point of that flag, and
which makes the address wrong rather than merely untidy.

`ruff check .` HAD NEVER PASSED — 105 errors. 93 were ruff's own safe fixes (mostly `datetime.UTC`
and import ordering); the other 12 needed judgement and are done by hand. Two are worth naming: a
dict on a dataclass is now `ClassVar`, which says out loud what `dataclass` was already inferring
from the missing annotation, and `int(round(x))` loses the cast because `round` already returns an
int. Both suites pass after the sweep, which is what makes a mechanical 39-file diff checkable.

AND ONE OF MY OWN, surfaced by the same mechanism. `test_the_dashboard_stays_crlf_and_two_space_json`
built its expected bytes by replacing every newline with a CRLF pair, so it asserted the file was
CRLF ON DISK. That is not a property of the repository: git stores the blob with LF, and what lands
in a working tree depends on the checker-out's `core.autocrlf`. It passed on the machine it was
written on and would have failed on every runner. It only became visible locally after a branch
switch re-materialised the file through git's own normalisation. It now compares normalised content
AND asserts the endings are consistent — a half-converted file being the state that actually makes a
diff unreadable. Positive-controlled against a 4-space reformat and against mixed endings.

THE FOURTH FAILURE IS THE ACCOUNTS JOB, AND ITS CAUSE WAS A LIST I COPIED. I took the API job's slim
pip install, which omits torch deliberately. That suite never imports `main` — also deliberately,
because importing it pulls torch — so it never reaches the lifespan. This job SPAWNS
`uvicorn src.main:app`, and the lifespan's SGLang branch does `import openai as _openai` to build
the backend client. Without it the app raises at startup, never answers, and the harness reports
that it never became ready — which reads like a broken harness rather than a missing dependency.
`openai` is in the list now.

Its failure is also legible from outside now: job logs need admin rights to download through the
API, so the first time this failed the only readable thing was a one-line annotation. It writes the
harness's own output and the last sixty lines into `$GITHUB_STEP_SUMMARY`.

Verified: `ruff check .` passes, root `pytest tests` exit 0, `pytest apps/api/tests` exit 0 FROM THE
REPOSITORY ROOT as CI runs it, desktop 2305 passed across 141 files, and all 47 shell steps across
the four workflows pass `bash -n`.

+++ scripts/bench_serving.py
@@ -148,7 +148,7 @@ async def one_request(
             usage = data.get("usage") or {}
             completion_tokens = usage.get("completion_tokens", 0)
             prompt_tokens = usage.get("prompt_tokens", 0)
-    except Exception as exc:  # noqa: BLE001 - every failure mode is a datum here
+    except Exception as exc:
         return RequestResult(
             ok=False,
             slot_seconds=time.monotonic() - start,
@@ -203,7 +203,7 @@ async def run_level(
     def percentile(values: list[float], q: float) -> float:
         if not values:
             return 0.0
-        index = min(len(values) - 1, int(round(q * (len(values) - 1))))
+        index = min(len(values) - 1, round(q * (len(values) - 1)))
         return values[index]
 
     return Level(
